Refactoring
- Added methods getInlineStyleForSelection, getCommonInlineStyleForSelection, getUnionInlineStyleForSelection to ContentState
- Fix: using inlineStyleOverride only for collapsed selection in EditorState
