+++ src/model/immutable/ContentState.js
@@ -18,6 +18,7 @@ import type {BlockNodeRecord} from 'BlockNodeRecord';
 import type DraftEntityInstance from 'DraftEntityInstance';
 import type {DraftEntityMutability} from 'DraftEntityMutability';
 import type {DraftEntityType} from 'DraftEntityType';
+import type {DraftInlineStyle} from 'DraftInlineStyle';
 
 const BlockMapBuilder = require('BlockMapBuilder');
 const CharacterMetadata = require('CharacterMetadata');
@@ -31,7 +32,7 @@ const SelectionState = require('SelectionState');
 const generateRandomKey = require('generateRandomKey');
 const sanitizeDraftText = require('sanitizeDraftText');
 
-const {List, Record, Repeat} = Immutable;
+const {List, Record, Repeat, OrderedSet} = Immutable;
 
 const experimentalTreeDataSupport = DraftFeatureFlags.draft_tree_data_support;
 
@@ -171,6 +172,40 @@ class ContentState extends ContentStateRecord {
     return DraftEntity.__get(key);
   }
 
+  getInlineStyleForSelection(selection: SelectionState): DraftInlineStyle {
+    if (selection.isCollapsed()) {
+      return getInlineStyleForCollapsedSelection(this, selection);
+    } else {
+      return getInlineStyleForNonCollapsedSelection(this, selection);
+    }
+  }
+
+  getCommonInlineStyleForSelection(
+    selection: SelectionState,
+  ): DraftInlineStyle {
+    if (selection.isCollapsed()) {
+      return getInlineStyleForCollapsedSelection(this, selection);
+    } else {
+      const {commonStyles} = getInlineStyleForNonCollapsedSelectionExt(
+        this,
+        selection,
+      );
+      return commonStyles;
+    }
+  }
+
+  getUnionInlineStyleForSelection(selection: SelectionState): DraftInlineStyle {
+    if (selection.isCollapsed()) {
+      return getInlineStyleForCollapsedSelection(this, selection);
+    } else {
+      const {foundStyles} = getInlineStyleForNonCollapsedSelectionExt(
+        this,
+        selection,
+      );
+      return foundStyles;
+    }
+  }
+
   static createFromBlockArray(
     // TODO: update flow type when we completely deprecate the old entity API
     blocks: Array<BlockNodeRecord> | {contentBlocks: Array<BlockNodeRecord>},
@@ -206,6 +241,142 @@ class ContentState extends ContentStateRecord {
     });
     return ContentState.createFromBlockArray(blocks);
   }
+}
+
+function getInlineStyleForCollapsedSelection(
+  content: ContentState,
+  selection: SelectionState,
+): DraftInlineStyle {
+  var startKey = selection.getStartKey();
+  var startOffset = selection.getStartOffset();
+  var startBlock = content.getBlockForKey(startKey);
+
+  // If the cursor is not at the start of the block, look backward to
+  // preserve the style of the preceding character.
+  if (startOffset > 0) {
+    return startBlock.getInlineStyleAt(startOffset - 1);
+  }
+
+  // The caret is at position zero in this block. If the block has any
+  // text at all, use the style of the first character.
+  if (startBlock.getLength()) {
+    return startBlock.getInlineStyleAt(0);
+  }
+
+  // Otherwise, look upward in the document to find the closest character.
+  return lookUpwardForInlineStyle(content, startKey);
+}
+
+function getInlineStyleForNonCollapsedSelection(
+  content: ContentState,
+  selection: SelectionState,
+): DraftInlineStyle {
+  var startKey = selection.getStartKey();
+  var startOffset = selection.getStartOffset();
+  var startBlock = content.getBlockForKey(startKey);
+
+  // If there is a character just inside the selection, use its style.
+  if (startOffset < startBlock.getLength()) {
+    return startBlock.getInlineStyleAt(startOffset);
+  }
+
+  // Check if the selection at the end of a non-empty block. Use the last
+  // style in the block.
+  if (startOffset > 0) {
+    return startBlock.getInlineStyleAt(startOffset - 1);
+  }
+
+  // Otherwise, look upward in the document to find the closest character.
+  return lookUpwardForInlineStyle(content, startKey);
+}
+
+function lookUpwardForInlineStyle(
+  content: ContentState,
+  fromKey: string,
+): DraftInlineStyle {
+  var lastNonEmpty = content
+    .getBlockMap()
+    .reverse()
+    .skipUntil((_, k) => k === fromKey)
+    .skip(1)
+    .skipUntil((block, _) => block.getLength())
+    .first();
+
+  if (lastNonEmpty)
+    return lastNonEmpty.getInlineStyleAt(lastNonEmpty.getLength() - 1);
+  return OrderedSet();
+}
+
+function getInlineStyleForNonCollapsedSelectionExt(
+  content: ContentState,
+  selection: SelectionState,
+): {commonStyles: DraftInlineStyle, foundStyles: DraftInlineStyle} {
+  let commonStyles = OrderedSet();
+  let foundStyles = OrderedSet();
+  const selStart = selection.getStartOffset();
+  const selEnd = selection.getEndOffset();
+  if (!selection.isCollapsed() && selStart >= 0 && selEnd >= 0) {
+    const blockMap = content.getBlockMap();
+    const blockKeys = Array.from(blockMap.keys());
+    let blocksByKeys = {};
+    blockMap.map(blk => {
+      blocksByKeys[blk.key] = blk;
+    });
+    const selStartKeyInd = blockKeys.indexOf(selection.getStartKey());
+    const selEndKeyInd = blockKeys.indexOf(selection.getEndKey());
+
+    let stylesLengths = {};
+    let selFullLength = 0;
+    for (let ind = selStartKeyInd; ind <= selEndKeyInd; ind++) {
+      const blockKey = blockKeys[ind];
+      const block = blocksByKeys[blockKey];
+
+      const chars = block.getCharacterList();
+      const blockCharsCount = chars.count();
+      let bsStart = 0;
+      let bsEnd = blockCharsCount;
+      if (ind == selStartKeyInd) bsStart = selStart;
+      if (ind == selEndKeyInd) bsEnd = selEnd;
+      const bsEndR = bsEnd - 1; //tip: 'R' means include last char, not exclude
+      const bsLength = bsEnd - bsStart;
+      selFullLength += bsLength;
+
+      block.findStyleRanges(
+        character => {
+          const styleKeysSet = character.getStyle(); //OrderedSet
+          return styleKeysSet !== null && styleKeysSet.size > 0;
+        },
+        (start, end) => {
+          const endR = end - 1; //tip: 'R' means include last char, not exclude
+          const isInSelection = !(
+            (start < bsStart && endR < bsStart) ||
+            (start > bsEndR && endR > bsEndR)
+          );
+          if (!isInSelection) return;
+          const fStart = Math.max(start, bsStart);
+          const fEnd = Math.min(end, bsEnd);
+          //const fEndR = fEnd - 1;
+          const len = fEnd - fStart;
+          const stlSet = block.getInlineStyleAt(start);
+          for (const stl of stlSet) {
+            foundStyles = foundStyles.add(stl);
+            if (!stylesLengths[stl]) stylesLengths[stl] = 0;
+            stylesLengths[stl] += len;
+          }
+        },
+      );
+    }
+
+    const commonStylesArr = Object.keys(stylesLengths).filter(
+      stl => stylesLengths[stl] == selFullLength,
+    );
+    commonStyles = OrderedSet(commonStylesArr);
+  }
+
+  return {
+    commonStyles,
+    foundStyles,
+  };
 }
 
 module.exports = ContentState;
